+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Release date `UNRELEASED`
 - Use `exchange.getRequestPath` instead of `.getRequestURI` which can contains
   host.
 - Accept `HttpHandler` as response instead of handler in the `:body`.
+- Upgrade dependencies - strojure/undertow "1.0.72-rc2", zmap "1.2.11".
 
 ## `1.0.52-beta8`